energy_agent/requirements.txt.txt
openai
python-dotenv
requests
duckduckgo-search
